Backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
PyMuPDF
requests
cloudinary
openpyxl
gunicorn